# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server API",
    "version": "1.0.0",
    "description": "Machine-readable API contract for the Express backend in server."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyAuthScheme": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative format: `Authorization: ApiKey <apiKey>`"
      },
      "inboundBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for worker ingress"
      }
    },
    "schemas": {
      "RequestId": {
        "type": "string"
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {},
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          }
        }
      },
...
      "EmailSender": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "nullable": true
              },
              "address": {
                "type": "string",
                "nullable": true
              }
            }
          },
          {
            "type": "null"
          }
        ]
      },
      "Email": {
        "type": "object",
        "description": "Full email including text/html body, returned by the detail route.",
        "required": [
          "id",
          "to",
          "localPart",
          "domain",
          "envelopeFrom",
          "from",
          "subject",
          "workerName",
          "sourceDomain",
          "messageId",
          "receivedAt",
          "createdAt",
          "hasRawMime",
          "rawMimeSize",
          "text",
          "html",
          "preview",
          "hasText",
          "hasHtml"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "to": {
            "type": "string"
          },
          "localPart": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "envelopeFrom": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "$ref": "#/components/schemas/EmailSender"
          },
          "subject": {
            "type": "string"
          },
          "workerName": {
            "type": "string"
          },
          "sourceDomain": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "hasRawMime": {
            "type": "boolean"
          },
          "rawMimeSize": {
            "type": "integer"
          },
          "text": {
            "type": "string"
          },
          "html": {
            "type": "string"
          },
          "preview": {
            "type": "string",
            "description": "First 400 characters of the text body, used to render list rows"
          },
          "hasText": {
            "type": "boolean"
          },
          "hasHtml": {
            "type": "boolean"
          },
          "rawMime": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "EmailRegister": {
        "type": "object",
        "required": [
          "id",
          "ownerUserId",
          "owner",
          "emailAddress",
          "localPart",
          "domain",
          "emailCount",
          "latestReceivedAt",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "ownerUserId": {
            "type": "integer"
          },
          "owner": {
            "$ref": "#/components/schemas/GroupOwner"
          },
          "emailAddress": {
            "type": "string"
          },
          "localPart": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "emailCount": {
            "type": "integer"
          },
          "latestReceivedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GroupOwner": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "id",
              "username",
              "displayName"
            ],
            "properties": {
              "id": {
                "type": "integer"
              },
              "username": {
                "type": "string"
              },
              "displayName": {
                "type": "string"
              }
            }
          },
          {
            "type": "null"
          }
        ]
      },
...
      "EmailRegisterCreateRequest": {
        "type": "object",
        "required": [
          "emailAddress"
        ],
        "properties": {
          "emailAddress": {
            "type": "string"
          }
        }
      },
      "SimpleSuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "requestId"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          }
        }
      },
...
      "TelegramPublicSettings": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "publicBaseUrl": {
            "type": "string"
          },
          "botTokenConfigured": {
            "type": "boolean"
          },
          "botTokenMasked": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TelegramOutboxStats": {
        "type": "object",
        "properties": {
          "pending": {
            "type": "integer"
          },
          "sending": {
            "type": "integer"
          },
          "sent": {
            "type": "integer"
          },
          "failed": {
            "type": "integer"
          }
        }
      },
      "TelegramRuntimeStatus": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "settings": {
            "$ref": "#/components/schemas/TelegramPublicSettings"
          },
          "webhookUrl": {
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastWebhookRegisteredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "workerActive": {
            "type": "boolean"
          },
          "processing": {
            "type": "boolean"
          },
          "lastPollAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastDeliveryAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastError": {
            "type": "string",
            "nullable": true
          },
          "lastResult": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "outbox": {
            "$ref": "#/components/schemas/TelegramOutboxStats"
          }
        }
      },
      "TelegramBotCommand": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "TelegramCommandsRegisterResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "count": {
            "type": "integer"
          },
          "commands": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TelegramBotCommand"
            }
          },
          "runtime": {
            "$ref": "#/components/schemas/TelegramRuntimeStatus"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          }
        }
      },
...
      "BlockedSenderMatch": {
        "type": "object",
        "description": "Present only when the inbound email was rejected by a blocked sender rule",
        "properties": {
          "id": {
            "type": "integer"
          },
          "patternType": {
            "type": "string",
            "enum": [
              "email",
              "domain"
            ]
          },
          "pattern": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "global",
              "domain"
            ]
          },
          "reason": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          }
        }
      },
      "EmailSummary": {
        "type": "object",
        "description": "Trimmed version used for lists; full text/html omitted so the payload does not grow with email count. Call GET /v1/emails/{id} to fetch the full content.",
        "required": [
          "id",
          "to",
          "localPart",
          "domain",
          "envelopeFrom",
          "from",
          "subject",
          "workerName",
          "sourceDomain",
          "messageId",
          "receivedAt",
          "createdAt",
          "hasRawMime",
          "rawMimeSize",
          "preview",
          "hasText",
          "hasHtml"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "to": {
            "type": "string"
          },
          "localPart": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "envelopeFrom": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "$ref": "#/components/schemas/EmailSender"
          },
          "subject": {
            "type": "string"
          },
          "workerName": {
            "type": "string"
          },
          "sourceDomain": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "hasRawMime": {
            "type": "boolean"
          },
          "rawMimeSize": {
            "type": "integer"
          },
          "preview": {
            "type": "string",
            "description": "First 400 characters of the text body, used to render list rows"
          },
          "hasText": {
            "type": "boolean"
          },
          "hasHtml": {
            "type": "boolean"
          }
        }
      }
    },
    "responses": {
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/email-registers": {
      "get": {
        "tags": [
          "EmailRegisters"
        ],
        "summary": "List mailbox registrations for the current user",
        "description": "Normal users only see their own registrations. Global admins may pass ownerUserId to inspect another user.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ownerUserId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "search",
            "description": "Case-insensitive substring filter on recipient_address; also narrows total.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registration list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "registrations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmailRegister"
                      }
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EmailRegisters"
        ],
        "summary": "Register a mailbox for monitoring",
        "description": "Mailbox must still be covered by the caller's permission scope. Each mailbox can only be registered by one user globally. Re-registering the same mailbox by the same user is idempotent.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRegisterCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mailbox registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "registration": {
                      "$ref": "#/components/schemas/EmailRegister"
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/v1/email-registers/new-mail": {
      "get": {
        "tags": [
          "EmailRegisters"
        ],
        "summary": "Generate and register a realistic new mailbox",
        "description": "Creates a realistic-looking mailbox address on an active domain available to the caller, then registers it immediately. Normal users are limited to domains where they have active permission. Global admins may use any active domain and may optionally target another owner with ownerUserId. This route accepts bearer session tokens and API keys.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "domain",
            "schema": {
              "type": "string"
            },
            "description": "Optional active domain to force mailbox generation onto. The caller must be allowed to use it; global admins may use any active domain."
          },
          {
            "in": "query",
            "name": "ownerUserId",
            "schema": {
              "type": "integer"
            },
            "description": "Optional owner override for global admins."
          }
        ],
        "responses": {
          "200": {
            "description": "Mailbox generated and registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "registration": {
                      "$ref": "#/components/schemas/EmailRegister"
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/v1/email-registers/{registrationId}": {
      "delete": {
        "tags": [
          "EmailRegisters"
        ],
        "summary": "Delete a mailbox registration",
        "description": "Owners may delete their own registration. Global admins may delete any registration.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "registrationId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registration deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleSuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/emails": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "List emails",
        "description": "Global admins can list all emails with scope=system. Any caller can force scope=registered to only see emails whose mailbox is both permitted and registered in email_registers for that user. domain/address/search filters narrow the result set. search supports multiple terms and checks subject, body, stored metadata and raw_mime when available.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string",
              "enum": [
                "registered",
                "system"
              ]
            }
          },
          {
            "in": "query",
            "name": "domain",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "address",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "emails": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmailSummary"
                      }
                    },
                    "hasMore": {
                      "type": "boolean"
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad query or missing scope filters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/emails/{id}": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get email by ID",
        "description": "Normal users can only read emails whose mailbox is registered for them and still covered by permission. Global admins can read any email.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "includeRawMime",
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "$ref": "#/components/schemas/Email"
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Emails"
        ],
        "summary": "Delete email by ID",
        "description": "Normal users must have domain permission for the mailbox domain and the mailbox must be registered for them.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleSuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/inbound/email": {
      "post": {
        "tags": [
          "Inbound"
        ],
        "summary": "Ingest raw MIME email from worker",
        "security": [
          {
            "inboundBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "message/rfc822": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-Email-Envelope-To",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Envelope-From",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Worker-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Received-At",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Domain",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Message-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Email accepted; blocked senders are acknowledged but discarded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "integer",
                      "nullable": true,
                      "description": "null when the email was blocked and therefore not stored"
                    },
                    "envelopeTo": {
                      "type": "string"
                    },
                    "domain": {
                      "type": "string"
                    },
                    "receivedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    },
                    "blocked": {
                      "type": "boolean",
                      "description": "true when a blocked sender rule discarded the email"
                    },
                    "blockedBy": {
                      "$ref": "#/components/schemas/BlockedSenderMatch"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/system/telegram/commands/register": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Register Telegram bot commands",
        "description": "Global admin only",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "responses": {
          "200": {
            "description": "Telegram commands registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramCommandsRegisterResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}